ISSUE-10 : moves all logic out of main.go to packages
